mooc_src/requirements.txt
Jinja2 ~= 3.1.6
PyYAML ~= 6.0.2
